Update Google API urls
These updates ensure we're using the latest API urls from Google and continue to provide a working app after the shutdown of the Google+ functionality in March 2019.

By changing to the new v3 url for the user info retrieval, some of the keys from the response have changed. I've added the mappings to the old keys for the raw user data so nothing should break. These can be removed in v5.0 of Socialite.