Create method inCallbackUri()
Create the method `inCallbackUri()` in order to be able to add parameters to the `callback_uri`
parameter of the redirection URL.

Similar to `with()`, but instead of appending the paramenters to the whole redirection URL,
they are appended to the `callback_uri` parameter.

As an example, imagine that we have the following data from the provider example.com:

```
CLIENT_ID=1234567890
CALLBACK_URI=http://myweb.com/oauth-callback
```

Generated redirection URL with `$socialite_provider->redirect()`:

```
https://provider.example.com/oauth/authorize
    ?client_id=123567890
    &redirect_uri=http%3A%2F%2Fmyweb.com%2Foauth-callback%3Fparam%3Dvalue
```

Now we want to send a key/value pair `param=value` and get it back.

A simple use case could be indicating the origin of the redirect: `from=login`, `from=profile`
or `from=manage-auths`...

Generated redirection URL with existing `$socialite_provider->with(['from' => 'login'])->redirect()`:

```
https://provider.example.com/oauth/authorize
    ?client_id=123567890
    &redirect_uri=http%3A%2F%2Fmyweb.com%2Foauth-callback
    &from=login
```

Here, `from=login` is passed to the provider, but is lost in the response.

Generated redirection URL with new `$socialite_provider->inCallbackUri(['from' => 'login'])->redirect()`:

```
https://provider.example.com/oauth/authorize
    ?client_id=123567890
    &redirect_uri=http%3A%2F%2Fmyweb.com%2Foauth-callback%3Ffrom%3Dlogin
```

Here, `from=login` is passed to the provider and sent back to us in the response!

The callback URL (parameter `redirect_uri` in the redirection URL), once urldecoded, is:

```
http://myweb.com/oauth-callback?from=login
```